backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
aiohttp==3.9.0
pycoingecko==3.1.0
PyGithub==2.1.1
tweepy==4.14.0
reportlab==4.0.7
matplotlib==3.8.2
pandas==2.1.3
numpy==1.26.2
pillow==10.1.0
python-multipart==0.0.6
jinja2==3.1.2
cachetools==5.3.2
